Add LectureList dynamic import
