+++ shared/common/src/components/index.ts
@@ -12,6 +12,7 @@ export { default as Header } from './Header'
 export { default as InquiryItem } from './InquiryItem'
 export { default as LectureApplyItem } from './LectureApplyItem'
 export { default as LectureItem } from './LectureItem'
+export { default as LectureList } from './LectureList'
 export { default as PaginationPages } from './PaginationPages'
 export { default as PostItem } from './PostItem'
 export { default as PrivateRouter } from './PrivateRouter'
@@ -20,4 +21,3 @@ export { default as Sequence } from './Sequence'
 export * from './Sliders'
 export { default as UserItem } from './UserItem'
 export { default as ValueInput } from './ValueInput'
-
